… go.sum
… vrf/secp256_k1_sha256_tai.json