(after apply PR #530)
```
benchstat before.txt after.txt                                                                                   
name          old time/op    new time/op    delta
VRF/prove-8      297µs ±10%     145µs ± 4%  -50.98%  (p=0.008 n=5+5)
VRF/verify-8     464µs ± 6%     223µs ± 4%  -51.89%  (p=0.008 n=5+5)

name          old alloc/op   new alloc/op   delta
VRF/prove-8     4.81kB ± 0%    3.67kB ± 3%  -23.63%  (p=0.008 n=5+5)
VRF/verify-8    4.94kB ± 0%    3.62kB ± 4%  -26.60%  (p=0.016 n=4+5)

name          old allocs/op  new allocs/op  delta
VRF/prove-8       93.0 ± 0%      74.8 ± 2%  -19.57%  (p=0.008 n=5+5)
VRF/verify-8       112 ± 0%        86 ± 2%  -23.21%  (p=0.008 n=5+5)
```